Arithmetic Operator
Arithmetic Operations
We can also perform mathematical operations using print.

print(2+4) # Output: 6
The above code will output the result of the mathematical operation, which is 6.

Note: If you enclose the operation within quotes, the entire expression will be treated as text (a string) and will be printed exactly as it is.

print("2+4") # Output: 2+4
In this case, the output will be the string "2 + 4" instead of the result of the addition.

Without quotes: The expression is evaluated, and the result is printed.
With quotes: The expression is treated as a string and printed as it is.
Understanding this distinction is important for correctly displaying the results of calculations versus showing text.