Variable in Python
Variables in Python
A variable is like a labelled box where you can store data. Imagine you have a box labeled "age" and you put the number 25 in it. In Python, you would do this by writing:

age = 25
When you write this, python creates a box(variable) with name
a
g
e
age and stores 25 in that box(variable).

This process of creating a variable to store a value is called Declaration and
The process of setting its value for the first time is called Initialization.
Here's the cool part: whenever you use age in your code, Python will remember it is 25. For example, if you write print(age), Python will show 25.

print(age) # Output: 25