Strings and Integers
Strings and Integers
Python uses the + sign for both addition and concatenation.

Numbers are added.
Strings are concatenated.
We cannot mix the two
Task
You are given a program which does the following

You want to output 2569
Try running the given code given in the IDE as it is - it will give a Compilation error
Fix the error so that both the variables become strings and the output is 2569