String length
String length
We can use the len() function to get the length of a string.

For eg.

greeting = "hello"
len_string = len(greeting)
# len_string now stores the value of the length of greeting
Task
Write a program which does the following

Create a variable txt and assign it the text NumeroTres
Use the len() function to output to the console the number of characters in txt