Taking User Input
So you understood from the video the different ways in which input can be read. Let's take another example - the following commands will be used to get the name and age from user.
The program then prints out the greeting using the variables
n
a
m
e
name and
a
g
e
age.

name = input()
print("Hello, " + name + "!")

age = input()
print("You are " + age + " years old.")
input() assumes that the input is a string.
You can convert it to an integer or numerical value using int()

age = int(input())
print("You are", age, "years old.")
Task
Write a program which does the following

Declare an integer variable num
Try taking a number from the console and assign it to num
Output num to the console